making sure all validate form feature works

+++ CHATBOX PROJECT/app/routes.py
@@ -47,6 +47,11 @@ def all_courses(ref):
         flash('User is not authenticated')
         return redirect(url_for('login'))
 
+    # check if the user url is correct
+    elif current_user.id != ref:
+        flash('Unauthorised access', 'error')
+        return redirect(url_for('home'))
+
     # Check if the user is a teacher
     elif current_user.has_role('teacher'):
         flash('User access invalid')
@@ -92,7 +97,7 @@ def all_courses(ref):
                 search_query = search_form.search.data
                 all_courses = models.Course.query.filter(models.Course.subject.like(f'%{search_query}%')).all()
                 if all_courses is None:
-                    flash("So sorry we can't find what you're looking for")
+                    flash("So sorry we can't find what you're looking for", 'error')
 
         # Render the page with the list of filtered courses
         return render_template('all_courses.html', form=form, search_form=search_form, all_courses=all_courses)
@@ -142,27 +147,25 @@ def create_acc():
 def create_acc_teacher():
     form = forms.register_teacher()
 
-    # If the user is already authenticated, redirect them to their courses
     if current_user.is_authenticated:
-        flash('User is logged in')
+        flash('User is already logged in')
         return redirect(url_for('my_courses', ref=current_user.id))
 
-    elif request.method == 'GET':
-        return render_template('create_acc_teacher.html', form=form, title='Create an account (teacher)')
+    if request.method == 'POST':
+        if form.validate_on_submit():  # This will check if the form submission is valid
+            new_user = models.User()
+            new_user.role = "teacher"
+            new_user.name = form.name.data
+            new_user.email = form.email.data
+            new_user.password = generate_password_hash(form.password.data)
+            db.session.add(new_user)
+            db.session.commit()
+            flash('Account created successfully! Please login with your new details.', 'success')
+            return redirect(url_for('login'))  # Redirect to login page for new user to login
+        else:
+            flash('Errors have occurred in form submission.', 'error')
 
-    else:
-        form.validate_on_submit()
-        new_user = models.User()
-        new_user.role = "teacher"
-        new_user.name = form.name.data
-        new_user.email = form.email.data
-        new_user.password = generate_password_hash(form.password.data)
-        db.session.add(new_user)
-        db.session.commit()
-        flash('Login with your details')
-        return redirect(url_for('my_courses', ref=new_user.id))
-
-
+    return render_template('create_acc_teacher.html', form=form, title='Create an account (teacher)')
 # Route for user login
 @app.route('/login', methods=['GET', 'POST'])
 def login():
@@ -170,13 +173,10 @@ def login():
 
     # If the user is already logged in, redirect to courses
     if current_user.is_authenticated:
-        flash('User is logged in')
+        flash('User is already logged in')
         return redirect(url_for('all_courses', ref=current_user.id))
 
-    elif request.method == 'GET':
-        return render_template('login.html', form=form)
-
-    elif request.method == 'POST':
+    if form.validate_on_submit():  # Validate the form when it's submitted
         email = form.email.data
         password = form.password.data
         user = models.User.query.filter_by(email=email).first()
@@ -189,13 +189,18 @@ def login():
         else:
             flash('Invalid credentials. Please try again.', 'error')
 
+    # GET request or if validation fails, re-render the form
     return render_template('login.html', form=form)
-
 
 # Route to display courses for the logged-in user
 @app.route("/my_courses/<int:ref>")
 def my_courses(ref):
-    if current_user.is_authenticated:
+    if current_user.id != ref:
+        flash('Unauthorised access', 'error')
+        return redirect(url_for('home'))
+
+    elif current_user.is_authenticated:
+
         courses = models.User.query.filter_by(id=ref).first_or_404()
         if courses is None:
             return redirect(url_for('all_courses', ref=current_user.id))
@@ -225,31 +230,39 @@ def not_found_error(error):
 def chats(ref, course_id):
     if not current_user.is_authenticated:
         return redirect(url_for('login'))
-    
+ 
+    # Ensure that the user is enrolled in the course
+    course = models.Course.query.get(course_id)
+    if not course or current_user not in course.people:
+        flash('unauthorised access to class', 'error')
+        return redirect(url_for('home'))
+
+    # Redirect to homepage if not authorized
+
     chats = models.Chat.query.filter_by(course_id=course_id).all()
+  
+    # Fetch the names of users participating in the chat
     name = []
-
-    # Fetching names of users participating in the chat
     for chat in chats:
-        user = models.User.query.filter_by(id=chat.person_id).all()
-        for name_user in user:
-            name.append(name_user.name)
-    
-    form = forms.enter_chat()
-    if request.method == "GET":
-        return render_template('chats.html', form=form, chats=chats, name=name, course_id=course_id)
+        user = models.User.query.filter_by(id=chat.user_id).first()
+        name.append(user.name) if user else name.append('Unknown User')
 
-    elif request.method == "POST":
-        form.validate_on_submit()
-        new_chat = models.Chat()
-        new_chat.course_id = course_id
-        new_chat.person_id = ref
-        new_chat.chat_content = form.chat.data
-        new_chat.chat_status = 0
-        new_chat.timestamp = datetime.now()
+    form = forms.enter_chat()
+ 
+    if form.validate_on_submit():  # Check if the form was submitted and valid
+        new_chat = models.Chat(
+            course_id=course_id,
+            user_id=ref,
+            chat_content=form.chat.data,
+            chat_status="unread",
+            timestamp=datetime.now()
+        )
         db.session.add(new_chat)
         db.session.commit()
+
         return redirect(url_for("chats", ref=ref, course_id=course_id))
+
+    return render_template('chats.html', form=form, chats=chats, name=name, course_id=course_id)
 
 
 # Route to create a new course (teachers only)
@@ -259,19 +272,13 @@ def create_course():
     if not current_user.is_authenticated:
         flash('User is not authenticated')
         return redirect(url_for('login'))
-   
+
     # Check if the user has a teacher role
-    elif current_user.has_role('teacher'):   
+    elif current_user.has_role('teacher'):
         form = forms.CreateCourseForm()
 
-        # Handle GET request - render the course creation form
-        if request.method == 'GET':
-            return render_template('create_course.html', form=form)
-        
-        # Handle POST request - form submission
-        else:
-            form.validate_on_submit()
-
+        # Handle form submission
+        if form.validate_on_submit():
             # Create a new course using data from the form
             new_course = models.Course(
                 joining_code=random.randint(1111, 9999),  # Generate a random joining code
@@ -283,31 +290,30 @@ def create_course():
             db.session.add(new_course)
             db.session.commit()
 
-            # Retrieve the user and the latest course created
+            # Retrieve the user and enroll in the latest created course
             user = models.User.query.get(current_user.id)
             last_course = models.Course.query.order_by(models.Course.id.desc()).first()
-            course = models.Course.query.get(last_course.id)
 
             # Ensure both the user and the course exist
-            if user and course:
-                # Check if the user is already enrolled in the course
-                if course in user.courses:
+            if user and last_course:
+                if last_course in user.courses:
                     flash('You are already enrolled in this course', 'warning')
-                    return redirect(url_for('my_courses', ref=current_user.id))
                 else:
                     # Add the course to the user's courses list
-                    user.courses.append(course)
+                    user.courses.append(last_course)
                     db.session.add(user)
                     db.session.commit()
-
                     flash('Course created successfully!', 'success')
-                    return redirect(url_for('my_courses', ref=current_user.id))  # Redirect to user's courses page
-    
+                return redirect(url_for('my_courses', ref=current_user.id))
+        else:
+            # If form validation fails, render the form with errors
+            flash('Form validation failed. Please correct the errors.', 'error')
+            return render_template('create_course.html', form=form)
+
     # If the user is a student, deny access to course creation
     else:
         flash('Student access unauthorized', 'error')
         return redirect(url_for('my_courses', ref=current_user.id))
-
 
 # Route to check if there are new messages in a course chat
 @app.route('/check_new_messages/<int:course_id>/<int:last_chat_id>', methods=['GET'])
@@ -327,9 +333,14 @@ def check_new_messages(course_id, last_chat_id):
 def delete_course(ref, course_id):
     # Check if the user is authenticated
     if not current_user.is_authenticated:
-        flash('User is not authenticated')
+        flash('User is not authenticated', 'error')
         return redirect(url_for('login'))
-    
+
+    # Check if the URL is coherent with the current user's ID
+    elif current_user.id != ref:
+        flash('Unauthorized access', 'error')
+        return redirect(url_for('home'))
+
     # Check if the user has a teacher role
     elif current_user.has_role('teacher'):
         # Retrieve the course to be deleted and the current user
@@ -338,35 +349,37 @@ def delete_course(ref, course_id):
 
         # Ensure both the user and the course exist
         if user and course_to_delete:
-            # Check if the course belongs to the user
+            # Check if the course belongs to the user (teacher)
             if course_to_delete in user.courses:
-                if course_to_delete:
-                    try:
-                        # Attempt to delete the course from the database
-                        db.session.delete(course_to_delete)
-                        db.session.commit()
-                        flash("Course has been deleted", 'success')
+                # Additional check to ensure the teacher is enrolled in the course
+                if course_to_delete in current_user.courses:
+                    if course_to_delete:
+                        try:
+                            # Attempt to delete the course from the database
+                            db.session.delete(course_to_delete)
+                            db.session.commit()
+                            flash("Course has been deleted", 'success')
+                            return redirect(url_for('my_courses', ref=ref))
+                        except Exception as e:
+                            # Handle any errors during deletion
+                            db.session.rollback()
+                            flash(f"Error deleting course: {e}", "warning")
+                            return redirect(url_for('my_courses', ref=ref))
+                    else:
+                        flash(f"No course found with ID {course_id}.", 'warning')
                         return redirect(url_for('my_courses', ref=ref))
-                    
-                    except Exception as e:
-                        # Handle any errors during deletion
-                        db.session.rollback()
-                        flash(f"Error deleting course: {e}", "warning")
-                        return redirect(url_for('my_courses', ref=ref))
-                
                 else:
-                    flash(f"No course found with ID {course_id}.", 'warning')
+                    flash("You are not enrolled in this course", 'error')
                     return redirect(url_for('my_courses', ref=ref))
-            
             else:
                 flash("Authorization invalid, this course is not yours", 'error')
                 return redirect(url_for('my_courses', ref=ref))
-        
         else:
             flash("User or course does not exist", 'error')
             return redirect(url_for('my_courses', ref=ref))
-    
+
     # If the user is a student, deny access to course deletion
     else:
         flash('Student access unauthorized', 'error')
-        return render_template('my_courses', ref=ref)
+        return redirect(url_for('my_courses', ref=ref))
+
